pyproject.toml
[tool.poetry]
name = "hal9001"
version = "1.3.3.7"
description = "Code for HAL9000 prop"
authors = [
  "Karl Fredrik Haugland <kfh@tla.wtf>",
]

[tool.poetry.dependencies]
python = "^3.12"
pygame = "^2.5.2"
mypy = "^1.10.0"
pydantic = "^2.7.1"
ruff = "^0.4.4"
pydantic-settings = "^2.2.1"
paho-mqtt = "^2.1.0"
google-cloud-texttospeech = "^2.16.3"
google-api-python-client = "^2.129.0"
oauth2client = "^4.1.3"
openai = "^1.30.1"
librosa = "^0.10.2.post1"

[tool.poetry.scripts]
hal9001 = "hal9001.cli:main"


[tool.ruff]
line-length = 120 # must be same as black
exclude = [
  ".git",
  ".github",
  ".venv",
  "__pycache__",
  "migrations",
  "node_modules",
  "temp-deploy",
]

[tool.ruff.lint]
select = ["C", "E", "F", "W", "B", "I"]

[tool.mypy]
allow_untyped_globals = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
exclude = ['\.venv', '.github']
files = '**/*.py'
follow_imports = "normal"
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
pretty = false # Set to "false" to handle bug in PyCharm's mypy tool
python_version = "3.11"
show_column_numbers = true
show_error_codes = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = false

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_untyped_fields = true

